Add a bunch of stuff
Add "Animate background" checkbox
Add "As list" animation list button to open window of animation list arranged in a table. Configuration of arrangement of this table needs smarter algo.
Spiff up the buttons a bit more to actually be mutually-exclusive
Create a window for the Animation List; when changing the active animation using this method, update the dropdown too
Stub out a thing to pass the press of a spiffy button so that we know what to act on
Add some more documentation
Make Exit func a bit more-descriptive. Watch out for that grue!
Add some appnames