Assemble death poses with dense packing
If the death pieces puzzle is solved slightly different it will become slightly more straight forward, and it can be packed in a dense manner which allows external file formats to save some space.

With this commit the lower part of the death sprite sheets will have the following layout in with the poses are numbered 0, 1, 3, 4, and pose 2 is composed of parts A, B, C (with the actual origin of the parts displayed to the right).

```
   0123456789ABCDEF
   . . . . . . . .
10 0001111444444444
11 0001111444444444
12 0001111444444444
13 0001111444444444
14 0001111444444444
15 0001111444444444
16 3333333444444444
17 3333333444444444
18 3333333444444444 AAAAA
19 3333333444444444 AAAAA
1A 3333333444444444 AAAAA
1B 3333333AAAAABBBB AAAAA
1C 3333333AAAAABBBB BBBBC
1D 3333333AAAAABBBB BBBBC
1D 3333333AAAAABBBB BBBBC
1E 3333333CCCC      BBBBC
```