Implement RDC file export for both Z3 and SM sprites
With this commit, the RDC export is prepared, but not yet wired up, to include an utf-8 encoded null terminated author name. An empty author name is assumed instead, resulting in a single null byte for this data field.

Note that the author name is supposed to correspond to the utf-16 null terminated "Author" field from the ZSPR format, not the ascii null terminated "Author rom display" field.