Implement Timed Palettes
I hate it but I had to resign to applying the palette variant (be it standard or otherwise) to every tile in a set of tiles if the variants differed within that set.
Ignore the frame_getter and just use the frame_number that was passed.
Replace the variant sent to the palette getter since it would just keep appending and might cause lag, dunno.