Encapsulation Refactor (#37)
* Added tests for garbage collection

* bugfix to library unit test

* BREAKING EVERYTHING

To fix memory leaks, had to comment out a lot of code...which means I have to add it all back in now.  *sad face*.  But the tests now pass!  #SmallVictories

* Travis CI fix

* Just work dammit

* Macs are apparently stupid

* Confirmed: Macs are stupid

* More Mac fixes

* Macs...

* Cleaning up YML

* Now Windows...

* Python 3.6 also, just in case

* Should work on newer Linux also

* many things are broken

but some are not

* Program loads again

But does not have most of its functionality...

* Finish merge from 'master'

* Test game garbage collection

* Attempt to fix plugins

Link's Z3R Sprite Download still isn't elegant

* Attempt to fix autodetect & spiffy panels

Add PNG dimensions to sprite manifest
Bring back spiffy panels
Pass resource_subpath to SpiffyFunctions
Read sprite manifests for PNG sizes and load accordingly
Use passed resource_subpath to make spiffy panels

* Added test to prevent PNGs from going to PhotoImage

* Sandboxed incompatible tests

* Travis CI path adjustment

* More Travis CI adjustments

* Travis...

* Move toward bettering StringVar support

* Make test output pretty

* stringvars renamed to spiffy_dict

* Fixed bug in background zooming

* Added test for zoom not crashing app

* Link displays but is not animated/transparent

* Samus displaying somewhat

* Moar tests/quick fix to fish

* Patch in some animatey bits

Link & Samus move, at one pose/frame, which is not ideal, but they move.
Currently set to not-animatey, but commenting out line #94 will make it animatey.
Line #97 is there in preparation for when the "frame_progression_table" is fixed.
Arrrrrrrrrrrrrrrr, matey.

* MOVE ALL THE THINGS

App Resources:
 App Icons
 App Names
 English Translation
 Bundled Game Backgrounds
 Spiffy Buttons Icons
 Animations Lists
 Input Layout Definitions
 Other Manifests
User Resources:
 Custom App Names
 Working Dirs
 Custom Game Backgrounds
 Custom Language Translations

This commit leaves it in a broken state. This commit is just moving stuff. The next commit(s) will be the code, as to maximize readability.

* Fix what I broke when I moved all the things

JSON files live in \manifests\
User-servicable stuff is in \user_resources\
Sprite Sheets (be they PNGs or ZSPRs or whatever) are in \sheets\
Resources exist in \app_resources\ (for App stuff) and \user_resources\ (for User stuff)
App Icons moved to \meta\icons\

* Fix autodetect

* Planet Zebes background

* Substantial increase in animateyness

* animation bugfixes.  aiming now works.

* Implement Timed Palettes

I hate it but I had to resign to applying the palette variant (be it standard or otherwise) to every tile in a set of tiles if the variants differed within that set.
Ignore the frame_getter and just use the frame_number that was passed.
Replace the variant sent to the palette getter since it would just keep appending and might cause lag, dunno.

* Shinespark Palette

Apply Shinespark Palette to Shinespark animation definitions

* Moar arrays

Switchers for Equipped Weapon and Control Codes
Switchers for Timed Palettes that are simple adjustments

* Add Link's Equipment

Snagged sheet from SpriteAnimator
Mathematically & programmatically slice main sheet into single prop images
Stuff into an associative array and save to disk for good measure
Implement as a plugin for testing

Does some work on #21

* Let user choose combo sprite to extract

It uses tk, sorry, but it works.

* Fix build script

Update credits to link to PixelArt as alluded to earlier.

* Transparency fixed, other bugfixes

* Fixed Overview and Reload

* Overview Scale Factor into Game Manifest

* Moar tests

* Condense more palettes

Death Suit
Death
Flash

* Use Death Suit palette

* Apply Screw Attack Palette

Applied palette to poses with one tile and suit tiles in poses with multiple tiles
Doesn't seem to actually apply when animated.

* Bugfix to screw attack palette

* fixed PNG export

* Update version number